+++ app/src/main/java/com/neo/fbrules/main/presenter/components/HistoricTextWatcher.kt
@@ -1,8 +1,9 @@
-package com.neo.fbrules.main.presenter.model
+package com.neo.fbrules.main.presenter.components
 
 import android.text.Editable
 import android.text.Selection
 import android.text.TextWatcher
+import com.neo.fbrules.main.presenter.model.HistoricModel
 import kotlinx.coroutines.*
 
 class HistoricTextWatcher(private val model: HistoricModel) : TextWatcher {
