+++ app/src/main/res/layout/content_text_rules_editor.xml
@@ -10,7 +10,7 @@
         android:layout_height="match_parent"
         android:fillViewport="true">
 
-        <com.neo.fbrules.main.presenter.view.TextRulesEditor
+        <com.neo.fbrules.main.presenter.components.view.TextRulesEditor
             android:id="@+id/text_rules_editor"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
