+++ app/src/main/res/layout/content_visual_rules_editor.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent">
 
-    <com.neo.fbrules.main.presenter.view.VisualRulesEditor
+    <com.neo.fbrules.main.presenter.components.view.VisualRulesEditor
         android:id="@+id/visual_rules_editor"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"/>
